mcp stats: address review findings and remove --genai CLI flag

+++ lib/MySQL_Thread.cpp
@@ -5387,8 +5387,8 @@ static bool mysql_pl_row_matches(const SQLite3_row* row, const processlist_query
 		return false;
 	}
 	if (opts.min_time_ms >= 0) {
-		const int row_time_ms = static_cast<int>(mysql_pl_to_u64(mysql_pl_field(row, mysql_processlist_columns_t::time_ms)));
-		if (row_time_ms < opts.min_time_ms) {
+		const uint64_t row_time_ms = mysql_pl_to_u64(mysql_pl_field(row, mysql_processlist_columns_t::time_ms));
+		if (row_time_ms < static_cast<uint64_t>(opts.min_time_ms)) {
 			return false;
 		}
 	}
